MonitorCard component created and Patch notifications method created
